build: reorganize sources
* remove external code
